Remove OrderHistoryTable, refactor OrderHistory
