+++ package.json
@@ -169,7 +169,6 @@
     "bfx-hf-server": "^3.5.0",
     "bfx-hf-strategy": "^1.0.1",
     "bfx-hf-util": "^1.0.1",
-    "ufx-ui": "git+https://github.com/bitfinexcom/ufx-ui",
     "bignumber.js": "^9.0.0",
     "bufferutil": "^4.0.3",
     "camelcase": "^5.3.1",
@@ -223,7 +222,7 @@
     "resolve": "^1.10.0",
     "sanitize-html": "^1.26.0",
     "thread-loader": "^2.1.2",
-    "ufx-ui": "git+https://github.com/bitfinexcom/ufx-ui.git#v0.1.0",
+    "ufx-ui": "git+https://github.com/bitfinexcom/ufx-ui#v0.1.0",
     "universal-analytics": "^0.4.20",
     "uuid": "^8.3.2",
     "workbox-webpack-plugin": "^5.0.0",
