--- src/components/OrderHistoryTable/OrderHistoryTable.props.js
@@ -1,13 +1,0 @@
-import PropTypes from 'prop-types'
-
-export const propTypes = {
-  onSelect: PropTypes.func,
-  onRemove: PropTypes.func,
-  orders: PropTypes.array,
-}
-
-export const defaultProps = {
-  orders: [],
-  onSelect: () => {},
-  onRemove: () => {},
-}
