--- src/components/OrderHistoryTable/index.js
@@ -1,3 +1,0 @@
-import OrderHistoryTable from './OrderHistoryTable.container'
-
-export default OrderHistoryTable
